move url params lines that use react-router-dom out of dashboards ViewDashboard to app ViewDashboard
