+++ ui/app/src/views/ViewDashboard.tsx
@@ -11,8 +11,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import { DashboardResource } from '@perses-dev/core';
+import { useState } from 'react';
+import { getUnixTime } from 'date-fns';
 import { ViewDashboard as DashboardView } from '@perses-dev/dashboards';
+import { TimeRangeStateProvider } from '@perses-dev/plugin-system';
+import { useSearchParams } from 'react-router-dom';
+import { DashboardResource, isDurationString, isRelativeValue, TimeRangeValue } from '@perses-dev/core';
 import { useSampleData } from '../utils/temp-sample-data';
 
 const DEFAULT_DASHBOARD_ID = 'node-exporter-full';
@@ -21,16 +25,56 @@ const DEFAULT_DASHBOARD_ID = 'node-exporter-full';
  * The View for viewing a Dashboard.
  */
 function ViewDashboard() {
-  const dashboard = useSampleData<DashboardResource>(
-    new URLSearchParams(window.location.search).get('dashboard') || DEFAULT_DASHBOARD_ID
-  );
+  const [searchParams, setSearchParams] = useSearchParams();
+
+  const dashboardParam = searchParams.get('dashboard');
+  const dashboard = useSampleData<DashboardResource>(dashboardParam || DEFAULT_DASHBOARD_ID);
+
+  const defaultDuration = dashboard?.spec.duration ?? '1h';
+
+  // TODO: preserve all existing params, change to use template variable approach
+  const fromParam = searchParams.get('from');
+  const toParam = searchParams.get('to');
+  const parsedParam = fromParam !== null ? fromParam.split('-')[1] : defaultDuration;
+  const pastDuration = parsedParam && isDurationString(parsedParam) ? parsedParam : defaultDuration;
+
+  // TODO: cleanup, fix types
+  const defaultTimeRange =
+    fromParam !== null && toParam !== null && toParam !== 'now'
+      ? { start: new Date(Number(fromParam)), end: new Date(Number(toParam)) }
+      : { pastDuration: pastDuration };
+  const [activeTimeRange, setActiveTimeRange] = useState<TimeRangeValue>(defaultTimeRange as TimeRangeValue);
 
   // TODO: Loading indicator
   if (dashboard === undefined) {
     return null;
   }
 
-  return <DashboardView dashboardResource={dashboard} />;
+  const handleOnDateRangeChange = (event: TimeRangeValue) => {
+    // TODO: create util to convert Perses RelativeTimeRange to GrafanaRelativeTimeRange (ex: from=now-1h&to=now)
+    if (isRelativeValue(event)) {
+      setSearchParams({
+        dashboard: dashboardParam ?? '',
+        from: `now-${event.pastDuration}`,
+        to: 'now',
+      });
+    } else {
+      const startUnixMs = getUnixTime(event.start) * 1000;
+      const endUnixMs = getUnixTime(event.end) * 1000;
+      setSearchParams({
+        dashboard: dashboardParam ?? '',
+        from: startUnixMs.toString(),
+        to: endUnixMs.toString(),
+      });
+      setActiveTimeRange({ start: event.start, end: event.end });
+    }
+  };
+
+  return (
+    <TimeRangeStateProvider initialValue={activeTimeRange} onDateRangeChange={handleOnDateRangeChange}>
+      <DashboardView dashboardResource={dashboard} />
+    </TimeRangeStateProvider>
+  );
 }
 
 export default ViewDashboard;
